requirements.txt
Django==4.2.3
PyDrive2==1.15.0
